Move saved objects loader to dashboard plugin (#124708)
* Move SavedObjectLoader into the dashboard plugin

* Remove old comment recommending savedobjectloader

* Fix types

(cherry picked from commit 00fb06cac5672954a945be27c395e9b3a2cfa5c2)
