+++ src/plugins/dashboard/public/services/string_utils.ts
@@ -6,13 +6,11 @@
  * Side Public License, v 1.
  */
 
-export class StringUtils {
-  /**
-   * Returns a version of the string with the first letter capitalized.
-   * @param str {string}
-   * @returns {string}
-   */
-  public static upperFirst(str: string = ''): string {
-    return str ? str.charAt(0).toUpperCase() + str.slice(1) : '';
-  }
+/**
+ * Returns a version of the string with the first letter capitalized.
+ * @param str {string}
+ * @returns {string}
+ */
+export function upperFirst(str: string = ''): string {
+  return str ? str.charAt(0).toUpperCase() + str.slice(1) : '';
 }
