+++ src/plugins/saved_objects/public/saved_object/index.ts
@@ -7,8 +7,6 @@
  */
 
 export { createSavedObjectClass } from './saved_object';
-export type { SavedObjectLoaderFindOptions } from './saved_object_loader';
-export { SavedObjectLoader } from './saved_object_loader';
 export { checkForDuplicateTitle } from './helpers/check_for_duplicate_title';
 export { saveWithConfirmation } from './helpers/save_with_confirmation';
 export { isErrorNonFatal } from './helpers/save_saved_object';
